Enable optional hiera usage.
This commit starts the work for plumbing hiera usage through the module.

The idea is that if servers have access to the hieradata then they will
get the nice things that hiera provides! If servers do not have access
the module will continue to work in a backwards compatible way while I
give access to those servers.