Add a britt role
This commit defines a britt role which can be applied to any node to add
my settings to it. Optionally if the developer parameter is set on this
new role all github repos created by the 'demophoon' name get checked
out into the '~/projects/' directory for the britt user account.