Fix retrieving keys from hash
This seems like a bug...

Basically what happens is whenever a function returns a hash the hash's
keys are symbolized. Even though Puppet still considers it a valid hash
none of the keys are accessible through the hash which is returned as
keys are looked up by string and are unable to be looked up by symbol.

To work around this i've implemented a function which performs the key
lookup on the symbolized hash.