:fire: Make https configurable in s3cfg
Note this should never be false, but because the version of ssl in s3cmd
isn't compatible with the servers version of ssl anymore we need to hack
our way around it.

Do not worry, the nodes affected are tunneling this traffic over an
internal VPN so this is really isn't as bad as it initially seems.