:lock: Reset ssh key for arcadia
This is really funny.. I ended up with a loaner laptop and my new laptop
came in the day after. Thanks to these configs I was able to get it up
and running in the state that it was in before in less time than it took
to wipe the loaner laptop.

Automation ftw!

Anyhow.. That's why I need to rotate out this key so soon.